+++ src/konpy/core.clj
@@ -2,8 +2,6 @@
   (:require [konpy.system :as system])
   (:gen-class))
 
-; (set! *default-data-reader-fn* tagged-literal)
-
 (defn -main
   [& _args]
   (system/start-system))