+++ src/konpy/utils.clj
@@ -45,3 +45,7 @@
       (str/replace #" " "")
       (str/replace #"\r\n" "")))
 
+(defn shorten
+  ([s] (shorten s 40))
+  ([s n] (let [pat (re-pattern (str "(^.{" n "}).*"))]
+           (str/replace-first s pat "$1..."))))